+++ mapbox-maps-ios/Sources/MapboxMaps/Annotations/AnnotationOrchestratorImpl.swift
@@ -164,7 +164,7 @@ internal final class AnnotationOrchestratorImpl: NSObject, AnnotationOrchestrato
 
                     // Get the identifiers of all the queried features
                     let queriedFeatureIds: [String] = queriedFeatures.compactMap {
-                        guard case let .string(featureId) = $0.feature.identifier else {
+                        guard case let .string(featureId) = $0.queriedFeature.feature.identifier else {
                             return nil
                         }
                         return featureId
@@ -187,7 +187,7 @@ internal final class AnnotationOrchestratorImpl: NSObject, AnnotationOrchestrato
 
         switch recognizer.state {
         case .began:
-            let layerIdentifiers = managers.map(\.layerId)
+            let layerIdentifiers = managers.flatMap(\.allLayerIds)
             let options = RenderedQueryOptions(layerIds: layerIdentifiers, filter: nil)
             let gestureLocation = recognizer.location(in: recognizer.view)
             mapFeatureQueryable.queryRenderedFeatures(with: gestureLocation, options: options) { result in
@@ -195,7 +195,7 @@ internal final class AnnotationOrchestratorImpl: NSObject, AnnotationOrchestrato
                 switch result {
                 case .success(let queriedFeatures):
                     let queriedFeatureIds: [String] = queriedFeatures.compactMap {
-                        guard case let .string(featureId) = $0.feature.identifier else {
+                        guard case let .string(featureId) = $0.queriedFeature.feature.identifier else {
                             return nil
                         }
                         return featureId