+++ mapbox-maps-ios/Tests/MapboxMapsTests/Gestures/GestureHandlers/AnyTouchGestureHandlerTests.swift
@@ -6,14 +6,16 @@ final class AnyTouchGestureHandlerTests: XCTestCase {
     var gestureRecognizer: MockGestureRecognizer!
     var cameraAnimatorsRunnerEnablable: MockMutableEnablable!
     var gestureHandler: AnyTouchGestureHandler!
+    var cameraAnimationsManager: MockCameraAnimationsManager!
 
     override func setUp() {
         super.setUp()
         gestureRecognizer = MockGestureRecognizer()
         cameraAnimatorsRunnerEnablable = MockMutableEnablable()
+        cameraAnimationsManager = MockCameraAnimationsManager()
         gestureHandler = AnyTouchGestureHandler(
             gestureRecognizer: gestureRecognizer,
-            cameraAnimatorsRunnerEnablable: cameraAnimatorsRunnerEnablable)
+            cameraAnimationsManager: cameraAnimationsManager)
     }
 
     override func tearDown() {
@@ -27,20 +29,20 @@ final class AnyTouchGestureHandlerTests: XCTestCase {
         gestureRecognizer.getStateStub.defaultReturnValue = .began
         gestureRecognizer.sendActions()
 
-        XCTAssertEqual(cameraAnimatorsRunnerEnablable.$isEnabled.setStub.invocations.map(\.parameters), [false])
+        XCTAssertEqual(cameraAnimationsManager.cancelAnimationsStub.invocations.count, 1)
     }
 
     func testGestureEnded() {
         gestureRecognizer.getStateStub.defaultReturnValue = .ended
         gestureRecognizer.sendActions()
 
-        XCTAssertEqual(cameraAnimatorsRunnerEnablable.$isEnabled.setStub.invocations.map(\.parameters), [true])
+        XCTAssertEqual(cameraAnimationsManager.cancelAnimationsStub.invocations.count, 0)
     }
 
     func testGestureCancelled() {
         gestureRecognizer.getStateStub.defaultReturnValue = .cancelled
         gestureRecognizer.sendActions()
 
-        XCTAssertEqual(cameraAnimatorsRunnerEnablable.$isEnabled.setStub.invocations.map(\.parameters), [true])
+        XCTAssertEqual(cameraAnimationsManager.cancelAnimationsStub.invocations.count, 0)
     }
 }